Move utils to lib
