+++ templates/skeleton/app/lib/variants.ts
